// src/redux/actions/action-types.ts
enum ActionType {
    ADD_NEW_LIST = "add_new_list",
    ADD_NEW_TASK = "add_new_task",
    DELETE_LIST = "delete_list",
    CHANGE_LIST_TITLE = "change_list_title",
    CHANGE_TASK_TITLE = "change_task_title",
    CHANGE_TASK_DESCRIPTION = "change_task_description",
    DELETE_TASK = "delete_task",
    CHANGE_LIST_OF_TASK = "change_list_of_task"
}

export default ActionType;